+++ src/features/records/hook/useRecordMutation.ts
@@ -1,0 +1,46 @@
+import { useMutation, useQueryClient } from "@tanstack/react-query";
+
+import { createRecord, updateRecord, updateDetail } from "../api/detailsApi";
+import { InterviewRecordCreateDTO, InterviewRecordUpdateDTO, RecordDetailUpdateDTO } from "../api/recordsDTOList";
+
+
+export function useRecordMutation() {
+    const queryclient = useQueryClient();
+    
+    const { mutateAsync: create } = useMutation({
+        mutationFn: ({ data }: { data: InterviewRecordCreateDTO }) =>
+            createRecord(data),
+        onSuccess: () => {
+            queryclient.refetchQueries({ queryKey: ["side"] });
+        },
+    });
+
+    const { mutate: update } = useMutation({
+        mutationFn: ({
+            recordId,
+            updatedata,
+        }: {
+            recordId: string;
+            updatedata: InterviewRecordUpdateDTO;
+        }) => updateRecord(recordId, updatedata),
+        onSuccess: () => {
+            const queryKeys = [["side"], ["record"]]; // 리패치할 쿼리 키들
+            queryKeys.forEach((key) => {
+                queryclient.refetchQueries({ queryKey: key });
+            });
+        },
+    });
+
+    const { mutate: updateDetailMutation } = useMutation({
+        mutationFn: ({
+            recordId,
+            index,
+            detail,
+        }: {
+            recordId: string;
+            index: number;
+            detail: RecordDetailUpdateDTO;
+        }) => updateDetail(recordId, index, detail),
+    })
+    return { create, update, updateDetailMutation }
+};
